requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
python-engineio==4.3.0
python-socketio==5.4.0
pyinstaller==5.13.0
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.3
websockets==11.0.3
pillow==10.0.0
